LightRAG/LightRAG-TagSystem-Demo/requirements.txt
# LightRAG标签体系Demo依赖包 - 用户系统版本
lightrag==0.1.0b6
openai>=1.97.1
chromadb==0.4.15
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2>=3.1.3
aiofiles==23.2.1
pandas==2.1.3
numpy>=1.24.0
networkx==3.2
flask==3.0.0
pyyaml==6.0.1
requests==2.31.0

# 用户系统新增依赖
pyjwt==2.8.0
cryptography>=41.0.7
bcrypt==4.0.1

# 开发工具
python-dotenv>=1.0.1

# 额外兼容性依赖
typing-extensions>=4.11
